Try to train the original LISTA network and ISTA algorithm.
The model can not converge due to the mismatch of the frequency of target and array, and also due to the two-paras of the network, which limits its ability to convergence.
For further work, 1. design a more complex network to learn the mismatch of frequency; 2. regenerate the dataset, which do not contain the mismatch data.